Aborted attempt at downloading files
Webvi library doesn’t seem to have a nice way of determining the file name
(“title”) without starting to download it. But doing that will means the file
has to be downloaded into a buffer if the user is verifying the file name, or
all other Webvi activity has to be blocked.